Introduce runWithErrorHandling function and move error handling logic there
